Small change to allow student score of 0: TODO - bug with object proerties
Only the values set by reading Strings have validation run on them

Noticed as the grading code would assign a score of 0, but I could not
manually put it in the property (it was incorrectly setup to require a min value of 1)

I vaugely remember maybe needing to side-step the validation for application internal code
for some reason (instantiating objects with defaults?) should probably just make sure
There is required to be a default value that passes validation for each proerty.

Should move this validation into the regular set() methods that take a java object
and call this from the parsing methods. This should hopefully reveal where I might have been
relying on thid behavior and it can be corrected.