TODO - define an interface for asking user for an expression, variable name and other information
This data is needed during an interactive action that can be performed on an Object.
The best current examples are in the ExpressionObject. Currently calls to create dialogs
appear in the ExpressionObject class, but this is meant to be a back-end abstraction
and really should have no GUI code in it. This can be refactored to communicate with
a genericized interface for collection data from users that could be passed in, or
created statically with a factory based on the particular GUI that should be used
given the current front-end driving the application.