Add docPanel null check in refactored PageGUI.
docPanel is used to find the currently selected object on a doc, in
which case the object may be drawn differently. The object drawing
method in PageGUI is sometimes used when there isn't an actual document
with a GUI representation (a docPanel). Currently it is used to redraw
an object in the background to find out its new bounds if something
changes like an expression or text that may expand the size of a text
box.