Fixed group creation
This bug was introduced by the addition of the overridden equals()
methods on MathObjects. The issue was an interaction between the new code
and two other behaviors. In the addObjectToPage() method there was a
check to make sure the object had not already been added. This doesn't
seem like the right behavior and might be hiding sloppy code elsewhere,
trying to remove this check however broke click and dragging to select 3
items. TODO - look back at this issue.

The code that ended up being modified to fix the bug was in the group()
method in NotebookPanel. After cloning the temporary group and adding
it to the page (which failed because the check to see that the object
was not already on the page with the new equality prevented the duplicated
object from being added) it would go ahead and delete the "old" copy which
was currently the temporary group. This seems to be fine, but I think the
behavior of adding an object described above seems wrong, if this is refactored
then the old strategy for the group creation could be brought back, but
I don't think this is necessary.

TODO - need to look at other parts of the code that may have broken due
to the new definition of equality on MathObjects, also should implement
hashCode() so that they can be used in Sets and HashMaps.