Fix bug in problem generation introduced by adding equals method to MathObjects
This was used to write unit tests, but this was at least one example
where the application was relying on the default object instance
equality definition. Would be good to come up with a strategy for finding
other hidden bugs that might have been introduced with this change.

Part of this bug was the anti-feature that is the guard against adding
an object to a page twice. This should probably produce an assert or
error rather than being silently ignored.